feat: migrate database config and update deployment

+++ app/models/cart.py
@@ -1,5 +1,5 @@
 """购物车模型"""
-from sqlalchemy import Column, DateTime, Float, ForeignKey, Integer, UniqueConstraint
+from sqlalchemy import Column, DateTime, ForeignKey, Integer, Numeric, UniqueConstraint
 from sqlalchemy.orm import relationship
 from sqlalchemy.sql import func
 
@@ -14,7 +14,7 @@ class CartItem(Base):
     user_id = Column(Integer, ForeignKey("users.id", ondelete="CASCADE"), nullable=False, index=True)
     product_id = Column(Integer, ForeignKey("products.id"), nullable=False)
     quantity = Column(Integer, nullable=False, default=1)
-    unit_price = Column(Float, nullable=False)  # 下单时的快照
+    unit_price = Column(Numeric(10, 2), nullable=False)  # 下单时的快照
     created_at = Column(DateTime, server_default=func.now())
 
     user = relationship("User", back_populates="cart_items")
