+++ app/models/order.py
@@ -1,5 +1,5 @@
 """订单模型"""
-from sqlalchemy import Column, DateTime, Float, ForeignKey, Integer, JSON, String
+from sqlalchemy import Column, DateTime, Float, ForeignKey, Integer, JSON, Numeric, String
 from sqlalchemy.orm import relationship
 from sqlalchemy.sql import func
 
@@ -12,10 +12,10 @@ class Order(Base):
     id = Column(Integer, primary_key=True, autoincrement=True)
     order_no = Column(String(32), unique=True, nullable=False, index=True)
     user_id = Column(Integer, ForeignKey("users.id"), nullable=False, index=True)
-    total_amount = Column(Float, nullable=False)
+    total_amount = Column(Numeric(10, 2), nullable=False)
     order_status = Column(String(32), nullable=False)   # PAID / PAY_FAILED / PENDING_PAYMENT
     payment_status = Column(String(32), nullable=False) # MOCK_SUCCESS / MOCK_FAILED / MOCK_PENDING
-    receiver = Column(JSON, nullable=False)             # {name, phone, address}
+    receiver = Column(JSON, nullable=False)            # {name, phone, address}
     remark = Column(String(200), nullable=True)
     created_at = Column(DateTime, server_default=func.now(), nullable=False)
 
@@ -31,8 +31,8 @@ class OrderItem(Base):
     product_id = Column(Integer, ForeignKey("products.id"), nullable=False)
     product_name = Column(String(100), nullable=False)  # 快照
     quantity = Column(Integer, nullable=False)
-    unit_price = Column(Float, nullable=False)
-    amount = Column(Float, nullable=False)
+    unit_price = Column(Numeric(10, 2), nullable=False)
+    amount = Column(Numeric(10, 2), nullable=False)
 
     order = relationship("Order", back_populates="items")
     product = relationship("Product")
